requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
alembic
asyncpg
pydantic-settings
sqlalchemy[asyncio]
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.135.1
google-auth==2.49.0
google-genai==1.66.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
passlib[bcrypt]
python-jose[cryptography]
python-multipart
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
starlette==0.52.1
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
websockets==16.0
